Added option to use total biomass in multi-stanza calculations
The original EwE6 code requires users to enter a biomass value
for the leading stanza group that is then used as the reference
value when calculating the remaining stanza group biomass values.
However, the calculation can just as easily be performed using
the total biomass across all sub-stanza groups as the reference,
and the latter value is sometimes more convenient.  The new code
offers this option in when filling in B and QB values for stanza
groups.