Adjusted flow matrix calculations to allow for floating point error
The calculation of the exported fraction of detritus and discard fate
can occassionally lead to negative values due to floating point
error in the summation of the other prescribed within-model fluxes.
I added a check for very small values in the calculation of these
values (tolerance = 1e-6), and zero out anything smaller than this.
My assumption is that no one will actually prescribe a fraction that
equates to 0.0001% of the detritus/discard fate.