Altered rpath2ecopathmodel to match read.rpath.params R syntax
In this update, I changed the input variables to the rpath2ecopathmodel
function to match the syntax of read.rpath.params function in the
Rpath package.  While this change is not backward compatible, it
makes this much more cross-compatible when importing and exporting
between R and Matlab, including synchronizing whch files are
required and which are optional.

I also made a few minor updates to the documentation of ecopathmodel2rpath
(which already matches write.rpath.params in syntax) to clarify
that the pedigree table is currently a placeholder in Rpath.